creating a suite file
