+++ test/operations_test/cascade_test/deletion_test.go
@@ -1,1 +1,1 @@
-package cascade
+package cascade_test
